feat: migrate sus command
